Resolved conflicts by inserting new features into original code files

+++ Scripts/PlayerScripts/PlayerVars.cs
@@ -5,7 +5,7 @@ public partial class PlayerVars : Node
 	public static PlayerVars Instance { get; private set; }
 
 	public Vector2 Position { get; set; }
-	public int Health { get; set; }
+	public float Health { get; set; }
 
 	public override void _Ready()
 	{
